Added test route for user and correct mistakes related to user creation and retrieving from a table. Initial create for dockerfile and docker-compose file

+++ Repositories/IUserRepository.cs
@@ -4,7 +4,7 @@ namespace PilotSchoolCheckIn.Repositories;
 
 public interface IUserRepository
 {
-	User GetById(long id);
+	User GetById(int id);
 	
 	void PostUser(User user);
 }
